+++ R/unnest.R
@@ -119,6 +119,10 @@ unnest_col <- function(.df, col = NULL, names_sep = NULL) {
 
   .l <- list_drop_empty(.l)
 
+  if (length(.l) == 0) {
+    .l <- list(logical())
+  }
+
   .check_data <- .l[[1]]
   is_vec <- is_simple_vector(.check_data)