+++ NEWS.md
@@ -13,6 +13,7 @@
 * Nested calls to `across.()` are handled properly (#505)
 * `across.()`: Can namespace functions in `.fns` arg (#511)
 * `as_tidytable()`: Can keep row names when converting a matrix (#527)
+* `unnest.()`: Handles empty data frames (@roboton, #530)
 
 # tidytable 0.8.0